ARROW-3698: [Gandiva] Segmentation fault when using a large table in Gandiva
https://issues.apache.org/jira/browse/ARROW-3698

Author: suquark <zsy9509@mail.ustc.edu.cn>

Closes #2902 from suquark/gandiva and squashes the following commits:

9c7ac4ac6 <suquark> typo
24904cb9f <suquark> lint
a1650aa7b <suquark> fix test
f76cb64e4 <suquark> fix types
72ef82197 <suquark> fix types
fdb777560 <suquark> lint
a30449fac <suquark> Add tests for selection vector.
cb4fc5de1 <suquark> Add test for huge filters.
1c3dd3562 <suquark> int64 selection vector support
89d2f5e9c <suquark> Add tests for huge table.
9f9928e7d <suquark> revert changes to ZSTD
0566a9cf0 <suquark> remote verbose
c51fb5c71 <suquark> fix include
7547de33d <suquark> Lint. Better error message.
44886728c <suquark> verbose
d005c2223 <suquark> fix typo
307411ec9 <suquark> retrigger ci
706ccaa29 <suquark> fix jni types
ca0338c99 <suquark> More fixes
4e4c8a574 <suquark> fix eval_batch.h & lint
207d79230 <suquark> Fix more possible integer overflows.
c4daa9945 <suquark> Fix integer overflow in Gandiva which causes segfault.

+++ cpp/src/gandiva/bitmap_accumulator.cc
@@ -22,7 +22,7 @@
 namespace gandiva {
 
 void BitMapAccumulator::ComputeResult(uint8_t* dst_bitmap) {
-  int num_records = eval_batch_.num_records();
+  int64_t num_records = eval_batch_.num_records();
 
   if (all_invalid_) {
     // set all bits to 0.
@@ -35,11 +35,11 @@ void BitMapAccumulator::ComputeResult(uint8_t* dst_bitmap) {
 /// Compute the intersection of multiple bitmaps.
 void BitMapAccumulator::IntersectBitMaps(uint8_t* dst_map,
                                          const std::vector<uint8_t*>& src_maps,
-                                         int num_records) {
+                                         int64_t num_records) {
   uint64_t* dst_map64 = reinterpret_cast<uint64_t*>(dst_map);
-  int num_words = (num_records + 63) / 64;  // aligned to 8-byte.
-  int num_bytes = num_words * 8;
-  int nmaps = static_cast<int>(src_maps.size());
+  int64_t num_words = (num_records + 63) / 64;  // aligned to 8-byte.
+  int64_t num_bytes = num_words * 8;
+  int64_t nmaps = src_maps.size();
 
   switch (nmaps) {
     case 0: {
@@ -58,7 +58,7 @@ void BitMapAccumulator::IntersectBitMaps(uint8_t* dst_map,
       // two src_maps bitmaps. do 64-bit ANDs
       uint64_t* src_maps0_64 = reinterpret_cast<uint64_t*>(src_maps[0]);
       uint64_t* src_maps1_64 = reinterpret_cast<uint64_t*>(src_maps[1]);
-      for (int i = 0; i < num_words; ++i) {
+      for (int64_t i = 0; i < num_words; ++i) {
         dst_map64[i] = src_maps0_64[i] & src_maps1_64[i];
       }
       break;
@@ -68,8 +68,8 @@ void BitMapAccumulator::IntersectBitMaps(uint8_t* dst_map,
       /* > 2 src_maps bitmaps. do 64-bit ANDs */
       uint64_t* src_maps0_64 = reinterpret_cast<uint64_t*>(src_maps[0]);
       memcpy(dst_map64, src_maps0_64, num_bytes);
-      for (int m = 1; m < nmaps; ++m) {
-        for (int i = 0; i < num_words; ++i) {
+      for (int64_t m = 1; m < nmaps; ++m) {
+        for (int64_t i = 0; i < num_words; ++i) {
           uint64_t* src_mapsm_64 = reinterpret_cast<uint64_t*>(src_maps[m]);
           dst_map64[i] &= src_mapsm_64[i];
         }
